# openapi/unified-openapi.yaml
openapi: 3.1.0
info:
  title: UserManagement
  version: 1.0.0
...
paths:
...
  /api/v1/auth/refresh:
    servers:
      - url: http://localhost:42185
        description: Generated server url
    post:
      tags:
        - Auth
      summary: Refresh access token
      operationId: refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserManagement_RefreshTokenRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserManagement_JsendResponseLoginResponse'
...
  /api/v1/auth/google-login:
    servers:
      - url: http://localhost:42185
        description: Generated server url
    post:
      tags:
        - Auth
      summary: Login or register with Google
      operationId: googleLogin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserManagement_GoogleLoginRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserManagement_JsendResponseLoginResponse'
  /api/v1/me:
    servers:
      - url: http://localhost:42185
        description: Generated server url
    get:
      tags:
        - Me
      summary: Get current user profile
      operationId: getMe
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserManagement_JsendResponseUserResponse'
    delete:
      tags:
        - Me
      summary: Delete current user account
      operationId: deleteMe
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: >-
                  #/components/schemas/UserManagement_JsendResponseDeleteAccountResponse
    patch:
      tags:
        - Me
      summary: Update current user profile
      operationId: patchMe
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserManagement_PatchUserRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserManagement_JsendResponseUserResponse'
  /api/v1/me/preferences:
    servers:
      - url: http://localhost:42185
        description: Generated server url
    get:
      tags:
        - Me
      summary: Get current user preferences
      operationId: getPreferences
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: >-
                  #/components/schemas/UserManagement_JsendResponseUserPreferencesResponse
    patch:
      tags:
        - Me
      summary: Update current user preferences
      operationId: patchPreferences
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: {}
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: >-
                  #/components/schemas/UserManagement_JsendResponseUserPreferencesResponse
...
components:
  schemas:
...
    UserManagement_RefreshTokenRequest:
      type: object
      properties:
        refreshToken:
          type: string
          minLength: 1
      required:
        - refreshToken
    UserManagement_JsendResponseLoginResponse:
      type: object
      properties:
        status:
          type: string
        data:
          $ref: '#/components/schemas/UserManagement_LoginResponse'
        message:
          type: string
    UserManagement_LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          format: int64
        preferences:
          $ref: '#/components/schemas/UserManagement_UserPreferencesResponse'
    UserManagement_UserPreferencesResponse:
      type: object
      properties:
        settings:
          type: object
          additionalProperties: {}
...
    UserManagement_GoogleLoginRequest:
      type: object
      properties:
        idToken:
          type: string
          minLength: 1
      required:
        - idToken
    UserManagement_JsonNullableMapStringObject:
      type: object
      properties:
        present:
          type: boolean
        undefined:
          type: boolean
    UserManagement_JsonNullableString:
      type: object
      properties:
        present:
          type: boolean
        undefined:
          type: boolean
    UserManagement_PatchUserRequest:
      type: object
      properties:
        fullName:
          $ref: '#/components/schemas/UserManagement_JsonNullableString'
        avatarUrl:
          $ref: '#/components/schemas/UserManagement_JsonNullableString'
        username:
          $ref: '#/components/schemas/UserManagement_JsonNullableString'
        preferences:
          $ref: '#/components/schemas/UserManagement_JsonNullableMapStringObject'
      required:
        - avatarUrl
        - fullName
        - username
    UserManagement_JsendResponseUserResponse:
      type: object
      properties:
        status:
          type: string
        data:
          $ref: '#/components/schemas/UserManagement_UserResponse'
        message:
          type: string
    UserManagement_UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        fullName:
          type: string
        username:
          type: string
        avatarUrl:
          type: string
        authProvider:
          type: string
        preferences:
          $ref: '#/components/schemas/UserManagement_UserPreferencesResponse'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserManagement_JsendResponseUserPreferencesResponse:
      type: object
      properties:
        status:
          type: string
        data:
          $ref: '#/components/schemas/UserManagement_UserPreferencesResponse'
        message:
          type: string
...
    UserManagement_DeleteAccountResponse:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        email:
          type: string
        fullName:
          type: string
        deletedAt:
          type: string
          format: date-time
    UserManagement_JsendResponseDeleteAccountResponse:
      type: object
      properties:
        status:
          type: string
        data:
          $ref: '#/components/schemas/UserManagement_DeleteAccountResponse'
        message:
          type: string
...