# services/user-management/src/main/java/io/github/phunguy65/zms/usermanagement/presentation/MeController.java
package io.github.phunguy65.zms.usermanagement.presentation;

import io.github.phunguy65.zms.shared.domain.Result;
import io.github.phunguy65.zms.shared.domain.valueobject.UserId;
import io.github.phunguy65.zms.shared.infrastructure.web.JsendResponse;
import io.github.phunguy65.zms.usermanagement.application.response.DeleteAccountResponse;
import io.github.phunguy65.zms.usermanagement.application.response.UserPreferencesResponse;
import io.github.phunguy65.zms.usermanagement.application.response.UserResponse;
import io.github.phunguy65.zms.usermanagement.application.usecase.DeleteAccountUseCase;
import io.github.phunguy65.zms.usermanagement.application.usecase.GetUserPreferencesUseCase;
import io.github.phunguy65.zms.usermanagement.application.usecase.GetUserUseCase;
import io.github.phunguy65.zms.usermanagement.application.usecase.PatchUpdatePreferencesUseCase;
import io.github.phunguy65.zms.usermanagement.application.usecase.PatchUpdateUserUseCase;
import io.github.phunguy65.zms.usermanagement.domain.AuthError;
import io.github.phunguy65.zms.usermanagement.presentation.request.PatchPreferencesRequest;
import io.github.phunguy65.zms.usermanagement.presentation.request.PatchUserRequest;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.validation.Valid;
import java.util.Map;
import java.util.UUID;
import org.openapitools.jackson.nullable.JsonNullable;
import org.springframework.http.ResponseEntity;
import org.springframework.security.core.Authentication;
import org.springframework.web.bind.annotation.*;

@RestController
@Tag(name = "Me", description = "Current user profile and preferences")
public class MeController extends BaseController {

    private final GetUserUseCase getUserUseCase;
    private final PatchUpdateUserUseCase patchUpdateUserUseCase;
    private final GetUserPreferencesUseCase getPreferencesUseCase;
    private final PatchUpdatePreferencesUseCase patchUpdatePreferencesUseCase;
    private final DeleteAccountUseCase deleteAccountUseCase;

    public MeController(
            GetUserUseCase getUserUseCase,
            PatchUpdateUserUseCase patchUpdateUserUseCase,
            GetUserPreferencesUseCase getPreferencesUseCase,
            PatchUpdatePreferencesUseCase patchUpdatePreferencesUseCase,
            DeleteAccountUseCase deleteAccountUseCase) {
        this.getUserUseCase = getUserUseCase;
        this.patchUpdateUserUseCase = patchUpdateUserUseCase;
        this.getPreferencesUseCase = getPreferencesUseCase;
        this.patchUpdatePreferencesUseCase = patchUpdatePreferencesUseCase;
        this.deleteAccountUseCase = deleteAccountUseCase;
    }

    @Operation(summary = "Get current user profile")
    @GetMapping(value = "/{version}/me", version = "1.0")
    public ResponseEntity<JsendResponse<UserResponse>> getMe(Authentication auth) {
        UserId userId = extractUserId(auth);
        if (userId == null) {
            return errorResponse(new AuthError.InvalidCredentials());
        }
        return switch (getUserUseCase.execute(userId)) {
            case Result.Success<UserResponse, AuthError> s ->
                ResponseEntity.ok(JsendResponse.success(s.value()));
            case Result.Failure<UserResponse, AuthError> f -> errorResponse(f.error());
        };
    }

    @Operation(summary = "Update current user profile")
    @PatchMapping(value = "/{version}/me", version = "1.0")
    public ResponseEntity<JsendResponse<UserResponse>> patchMe(
            @Valid @RequestBody PatchUserRequest dto, Authentication auth) {
        UserId userId = extractUserId(auth);
        if (userId == null) {
            return errorResponse(new AuthError.InvalidCredentials());
        }
        return switch (patchUpdateUserUseCase.execute(userId, dto.toCommand())) {
            case Result.Success<UserResponse, AuthError> s ->
                ResponseEntity.ok(JsendResponse.success(s.value()));
            case Result.Failure<UserResponse, AuthError> f -> errorResponse(f.error());
        };
    }

    @Operation(summary = "Get current user preferences")
    @GetMapping(value = "/{version}/me/preferences", version = "1.0")
    public ResponseEntity<JsendResponse<UserPreferencesResponse>> getPreferences(
            Authentication auth) {
        UserId userId = extractUserId(auth);
        if (userId == null) {
            return errorResponse(new AuthError.InvalidCredentials());
        }
        return switch (getPreferencesUseCase.execute(userId)) {
            case Result.Success<UserPreferencesResponse, AuthError> s ->
                ResponseEntity.ok(JsendResponse.success(s.value()));
            case Result.Failure<UserPreferencesResponse, AuthError> f -> errorResponse(f.error());
        };
    }

    @Operation(summary = "Update current user preferences")
    @PatchMapping(value = "/{version}/me/preferences", version = "1.0")
    public ResponseEntity<JsendResponse<UserPreferencesResponse>> patchPreferences(
            @Valid @RequestBody Map<String, Object> body, Authentication auth) {
        UserId userId = extractUserId(auth);
        if (userId == null) {
            return errorResponse(new AuthError.InvalidCredentials());
        }
        var dto = new PatchPreferencesRequest(JsonNullable.of(body));
        return switch (patchUpdatePreferencesUseCase.execute(userId, dto.toCommand())) {
            case Result.Success<UserPreferencesResponse, AuthError> s ->
                ResponseEntity.ok(JsendResponse.success(s.value()));
            case Result.Failure<UserPreferencesResponse, AuthError> f -> errorResponse(f.error());
        };
    }

    @Operation(summary = "Delete current user account")
    @DeleteMapping(value = "/{version}/me", version = "1.0")
    public ResponseEntity<JsendResponse<DeleteAccountResponse>> deleteMe(Authentication auth) {
        UserId userId = extractUserId(auth);
        if (userId == null) {
            return errorResponse(new AuthError.InvalidCredentials());
        }
        return switch (deleteAccountUseCase.execute(userId)) {
            case Result.Success<DeleteAccountResponse, AuthError> s ->
                ResponseEntity.ok(JsendResponse.success(s.value()));
            case Result.Failure<DeleteAccountResponse, AuthError> f -> errorResponse(f.error());
        };
    }

    private UserId extractUserId(Authentication auth) {
        if (auth == null || !(auth.getPrincipal() instanceof String principalId)) {
            return null;
        }
        try {
            return UserId.of(UUID.fromString(principalId));
        } catch (IllegalArgumentException e) {
            return null;
        }
    }
}